Move TypeScript writer to typescript/writer.ts subdirectory
