+++ src/api/index.ts
@@ -12,4 +12,4 @@ export { LogLevel } from "../utils/codegen-logger";
 export type { APIBuilderOptions, LocalStructureDefinitionConfig } from "./builder";
 export { APIBuilder, prettyReport } from "./builder";
 export type { CSharpGeneratorOptions } from "./writer-generator/csharp/csharp";
-export type { TypeScriptOptions } from "./writer-generator/typescript";
+export type { TypeScriptOptions } from "./writer-generator/typescript/writer";
